+++ src/components/GenericPagesComponents/Form/Form.ts
@@ -1,6 +1,6 @@
 import styled from "styled-components";
 
-import { box } from "../SharedComponents/Box";
+import { box } from "../../SharedComponents/Box";
 
 const Form = styled.form`
   ${box}
