+++ src/components/GenericPagesComponents/Form/index.ts
@@ -1,0 +1,1 @@
+export { default } from "./Form";
