+++ src/components/GenericPagesComponents/NewsletterForm/NewsletterForm.tsx
@@ -7,7 +7,7 @@ import Input from "@/components/SharedComponents/Input";
 import OakError from "@/errors/OakError";
 import DropdownSelect from "@/components/DropdownSelect";
 import errorReporter from "@/common-lib/error-reporter";
-import Form from "@/components/Forms/Form";
+import Form from "@/components/GenericPagesComponents/Form";
 import { BoxProps } from "@/components/SharedComponents/Box";
 import {
   USER_ROLES,
